refactor: merge whole `webui` folder into single file
