+++ internal/core/broker/vhost/admin.go
@@ -4,6 +4,16 @@ import (
 	"log"
 )
 
+
+
+const (
+	ADMIN_QUEUES      = "ottermq.admin.queues"
+	ADMIN_EXCHANGES   = "ottermq.admin.exchanges"
+	ADMIN_BINDINGS    = "ottermq.admin.bindings"
+	ADMIN_CONNECTIONS = "ottermq.admin.connections"
+)
+
+
 func (vh *VHost) bindAdminQueues() {
 	vh.mu.Lock()
 	defer vh.mu.Unlock()
