+++ web/handlers/api/queues.go
@@ -154,5 +154,4 @@ func GetMessage(c *fiber.Ctx, ch *amqp091.Channel) error {
 	return c.Status(fiber.StatusOK).JSON(fiber.Map{
 		"data": string(msg.Body),
 	})
-
 }
